refactor: split HLE bios module
This did grow way too big.
